+++ lib/src/widgets/badge/badge_label.dart
@@ -1,0 +1,11 @@
+// Copyright 2023 the Ako authors (https://pub.dev/packages/ako).
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+/// AkoBadgeLabel is a extendable class for [AkoBadge]
+class AkoBadgeLabel {
+  final String label;
+
+  AkoBadgeLabel(this.label);
+}